requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
beautifulsoup4==4.9.1
Brotli==1.0.7
certifi==2020.4.5.2
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
dash==1.10.0
dash-core-components==1.9.0
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
Django==3.0.7
django-filter==2.3.0
django-pandas==0.6.2
django-plotly-dash==1.3.1
django-redis==4.12.1
django-support==0.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dpd-components==0.1.0
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.1
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
msgpack==0.6.2
multitasking==0.0.9
numpy==1.18.5
pandas==0.25.3
pandas-datareader==0.8.1
plotly==4.8.1
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.23.0
retrying==1.3.3
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.1.0
yfinance==0.1.54
zipp==3.1.0
zope.interface==5.1.0
